+++ MelonLoader/Melons/RottenMelon.cs
@@ -10,16 +10,24 @@ namespace MelonLoader
         public readonly MelonAssembly assembly;
         public readonly Type type;
         public readonly string errorMessage;
-        public readonly Exception exception;
+        public readonly string exception;
 
         public RottenMelon(MelonAssembly assembly, string errorMessage, Exception exception = null)
         {
             this.assembly = assembly;
             this.errorMessage = errorMessage;
-            this.exception = exception;
+            this.exception = exception.ToString();
         }
 
         public RottenMelon(Type type, string errorMessage, Exception exception = null)
+        {
+            assembly = MelonAssembly.LoadMelonAssembly(null, type.Assembly);
+            this.type = type;
+            this.errorMessage = errorMessage;
+            this.exception = exception.ToString();
+        }
+
+        public RottenMelon(Type type, string errorMessage, string exception = null)
         {
             assembly = MelonAssembly.LoadMelonAssembly(null, type.Assembly);
             this.type = type;